backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
pydantic==2.4.2
python-multipart==0.0.6
PyJWT==2.8.0